# BodyDetection.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BodyDetection API",
    "description": "API for body detection, video processing, and image/video upload functionalities.",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/detect/images": {
      "post": {
        "summary": "检测图片",
        "description": "对上传的图片进行目标检测，返回检测结果。",
        "tags": [
          "Image"
        ],
        "requestBody": {
          "description": "上传需要检测的图片",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功返回检测结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detections": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "label": {
                            "type": "string",
                            "description": "检测到的目标类别"
                          },
                          "confidence": {
                            "type": "number",
                            "format": "float",
                            "description": "检测到的目标置信度"
                          },
                          "bbox": {
                            "type": "array",
                            "items": {
                              "type": "number"
                            },
                            "description": "检测到的目标边界框坐标 [x, y, width, height]"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/detect/videos": {
      "post": {
        "summary": "检测视频",
        "description": "对上传的视频进行目标检测，返回检测结果。",
        "tags": [
          "Video"
        ],
        "requestBody": {
          "description": "上传需要检测的视频",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "video": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功返回检测结果",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detections": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "timestamp": {
                            "type": "number",
                            "format": "float",
                            "description": "检测到的目标的时间戳（秒）"
                          },
                          "detections": {
                            "type": "array",
                            "items": {
                              "type": "object",
                              "properties": {
                                "label": {
                                  "type": "string",
                                  "description": "检测到的目标类别"
                                },
                                "confidence": {
                                  "type": "number",
                                  "format": "float",
                                  "description": "检测到的目标置信度"
                                },
                                "bbox": {
                                  "type": "array",
                                  "items": {
                                    "type": "number"
                                  },
                                  "description": "检测到的目标边界框坐标 [x, y, width, height]"
                                }
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}